backend/utils/general/get_env.py
import os
from pathlib import Path
from dotenv import load_dotenv

dotenv_path = Path(__file__).resolve().parent.parent.parent / ".env"
load_dotenv(dotenv_path=dotenv_path)
    
def getenv(key, default=None):
    return os.getenv(key, default)
