backend/requirements.txt
fastapi
uvicorn[standard]
redis
python-dotenv
pydantic
passlib[bcrypt]
python-jose[cryptography]
pytest
pytest-cov
python-multipart
requests
apscheduler
httpx
aiofiles
sqlalchemy
psycopg2-binary
google-generativeai
